requirements.txt
spidev
pyftdi